+++ src/App.tsx
@@ -6,12 +6,12 @@ import Home from "./pages/Home";
 import PastHackathons from "./pages/PastHackathons";
 import SponsorUs from "./pages/SponsorUs";
 import ContactUs from "./pages/ContactUs";
-import Application from "./pages/Application";
+import Application from "./pages/Archive/Application";
 // import Login from "./pages/Login";
 // import ResetPassword from "./pages/ResetPassword";
 import Error from "./pages/404";
-import DesignMerced from "./pages/DesignMerced";
-import HackMercedVI from "./pages/HackMercedVI";
+import DesignMerced from "./pages/Archive/DesignMerced";
+import HackMercedVI from "./pages/Archive/HackMercedVI";
 // import SignUp from "./pages/SignUp";
 import Maintenance from "./pages/Maintenance";
 import Dashboard from "./pages/Dashboard";
