+++ src/pages/Archive/HackMercedVI/styles.scss
@@ -12,7 +12,7 @@
         }
 
         &__container {
-            background-image: url("../../assets/images/landing-page-hmvi.png");
+            background-image: url("../../../assets/images/landing-page-hmvi.png");
             background-size: 100% 100%;
             background-repeat: no-repeat;
             height: 100%;
