+++ src/pages/Archive/ResetPassword/styles.scss
